+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 0.5.10 (SNAPSHOT)
-*
+* Add proxy support to updep target
 
 ### 0.5.9
 * Add invalid package.json and .bob.json error handling.